+++ modules/apps/commerce/commerce-inventory-web/src/main/resources/META-INF/resources/init.jsp
@@ -31,8 +31,8 @@ page import="com.liferay.commerce.inventory.model.CommerceInventoryReplenishment
 page import="com.liferay.commerce.inventory.model.CommerceInventoryWarehouse" %><%@
 page import="com.liferay.commerce.inventory.model.CommerceInventoryWarehouseItem" %><%@
 page import="com.liferay.commerce.inventory.web.internal.constants.CommerceInventoryFDSNames" %><%@
+page import="com.liferay.commerce.inventory.web.internal.constants.CommerceInventoryScreenNavigationConstants" %><%@
 page import="com.liferay.commerce.inventory.web.internal.display.context.CommerceInventoryDisplayContext" %><%@
-page import="com.liferay.commerce.inventory.web.internal.servlet.taglib.ui.constants.CommerceInventoryScreenNavigationConstants" %><%@
 page import="com.liferay.portal.kernel.language.LanguageUtil" %><%@
 page import="com.liferay.portal.kernel.util.CalendarFactoryUtil" %><%@
 page import="com.liferay.portal.kernel.util.Constants" %><%@
